feat: add hooks to fetch availability dimensions and values, refactor dynamic queries into dir
