+++ src/components/inputs/select/objectTypeSelect/objectTypeSelect.component.tsx
@@ -57,6 +57,7 @@ export const ObjectTypeSelect = forwardRef(
         options={options}
         ref={ref}
         onChange={onChangeWrapper}
+        withSearch
       />
     );
   },
