+++ src/enums/graphql.ts
@@ -2,6 +2,7 @@ export enum QueryKeys {
   Schema = "schema",
   ObjectTypesConfig = "objectTypesConfig",
   Search = "search",
+  AvailabilityDimensions = "availabilityDimensions",
   GetObject = "getObject",
   GetObjectAvailability = "getObjectAvailability",
   GetObjectRelationships = "getObjectRelationships",
