+++ src/interfaces/skylark/parsedObjects.ts
@@ -1,5 +1,7 @@
 import {
   SkylarkGraphQLAvailabilityDimension,
+  SkylarkGraphQLAvailabilityDimensionWithValues,
+  SkylarkGraphQLAvailabilityDimensionValue,
   SkylarkGraphQLObjectImage,
 } from "./gqlObjects";
 import {
@@ -23,6 +25,11 @@ export enum AvailabilityStatus {
   Unavailable = "Unavailable",
 }
 
+export interface ParsedSkylarkDimensionsWithValues
+  extends SkylarkGraphQLAvailabilityDimension {
+  values: SkylarkGraphQLAvailabilityDimensionValue[];
+}
+
 export interface ParsedSkylarkObjectAvailabilityObject {
   uid: SkylarkUID;
   external_id: SkylarkExternalId;
@@ -33,7 +40,7 @@ export interface ParsedSkylarkObjectAvailabilityObject {
   timezone: string;
   neverExpires: boolean;
   status: AvailabilityStatus | null;
-  dimensions: SkylarkGraphQLAvailabilityDimension[];
+  dimensions: SkylarkGraphQLAvailabilityDimensionWithValues[];
 }
 
 export interface ParsedSkylarkObjectAvailability {
